admin only stuff

+++ app/controllers/admin/content_controller.rb
@@ -114,21 +114,27 @@ class Admin::ContentController < Admin::BaseController
   end
 
   def merge
-    id1 = params['article_id'] 
-    id2 = params['merge_with'] 
-    if id1 && id2
-      article1 = Article.find(id1)  
-      article2 = Article.find(id2)
-      article1.update_attributes body: "#{article1.body} #{article2.body}"
-      article2.comments.each do |comment|
-        comment.update_attributes article_id: id1
+    if current_user.admin?
+      id1 = params['article_id'] 
+      id2 = params['merge_with'] 
+      if id1 && id2
+        article1 = Article.find(id1)  
+        article2 = Article.find(id2)
+        article1.update_attributes body: "#{article1.body} #{article2.body}"
+        article2.comments.each do |comment|
+          comment.update_attributes article_id: id1
+        end
+        article2.destroy
+        flash[:info] = _("Successfully merged")
+        redirect_to "/admin/content/edit/#{id1}"
+      else
+        flash[:error] = _("Error, didn't specify 2 articles to merge #{params.inspect}")
+        redirect_to "/admin/content/edit/#{id1}"
+        return
       end
-      article2.destroy
-      flash[:info] = _("Successfully merged")
-      redirect_to "/admin/content/edit/#{id1}"
     else
-      redirect_to "/admin/content/edit/#{id1}"
-      flash[:error] = _("Error, didn't specify 2 articles to merge #{params.inspect}")
+      flash[:error] = _("you need to be admin")
+      redirect_to :action => 'index'
       return
     end
   end
